game: prototype and out-param hygiene behind the pointer-width fixes
Follow-up tier to b88f05de, from the same -W sweep. One of these is not
hygiene at all:

- bgDebPrintROOMID returns a function-local static char[10][9] and had
  no prototype anywhere, so gnu89 gave it an implicit int return and
  callers truncated a STATIC's address - the exact save_data disease
  b88f05de fixed in file2.h, in its second hiding place. Prototype in
  bg.h; the receiving local in bondview2.c was also mistyped s32*.
- mp_respawn_handler declared its StandTile* as s32 (bondview_r.c's copy
  of the same code already had it right). Works today because the heap
  is kept low; typed honestly now.
- indycmdReceiveFile's third parameter is written through by
  indyrescmdResponseRecieveFile but was declared u32, and the caller
  squeezed &size through it. Dead in practice (osReadHost/osWriteHost
  are empty stubs and the ob.c path never triggers with a real ROM), but
  it round-trips as u32* now.
- prop.c and music.c now include port.h: portSwapSetupFile and the
  audio-bank promotions were called without prototypes and their
  returned pointers assigned to live globals through implicit int.
  Verified behavior-neutral by diffing normalized disassembly of both
  TUs - only the sign-extension instructions disappear.
- pi.c/vi.c passed &array where the element pointer is wanted (same
  address, wrong type).

Sweep delta across all 236 TUs: 0 new warnings, 21 removed. The 45
remaining unprototyped pointer returns were each traced to heap (memp,
mema, dyn, langGet, portLowAlloc) and left alone; the remaining
font-table s32 stores were traced to mempAllocBytesInBank allocations -
heap, safe on every target including Android PIE.

Verified: both golden suites bit-identical, PIE soak exit 0 with the
pinned pre-change STATEHASH (crc fd68d09f at tick 59000), arm64 suite
PASS, and the device runs the play harness clean.

Found and fixed by a review agent; audited and verified before commit.

+++ src/game/indy_comms.c
@@ -32,7 +32,7 @@ void indycommHostLoadFile(char *filename, u8 *targetloc)
     if (indy_ready)
     {
         indycmdSendLoadFile(filename,0x400000);
-        indycmdReceiveFile(response1,response2,(u32)&size,targetloc);
+        indycmdReceiveFile(response1,response2,&size,targetloc);
     }
     return;
 }
